feat: Add dynamic metadata generation function
